+++ lib/src/app.dart
@@ -6,6 +6,7 @@ import 'package:get/get.dart';
 import '/src/config/config.dart';
 import '/src/utils/binding.dart';
 import 'modules/locale/data/services/localization_service.dart';
+import 'modules/connections/connection.dart';
 
 class App extends StatelessWidget {
   const App({Key? key}) : super(key: key);
@@ -17,13 +18,15 @@ class App extends StatelessWidget {
         color: Colors.white,
         size: 50.0,
       ),
-      child: GetMaterialApp(
-        initialBinding: InitialBindings(),
-        getPages: RouteManager.instance.pages,
-        translations: LocalizationService(),
-        locale: LocalizationService.locale,
-        fallbackLocale: LocalizationService.fallbackLocale,
-        theme: theme,
+      child: ConnectionOverlay(
+        child: GetMaterialApp(
+          initialBinding: InitialBindings(),
+          getPages: RouteManager.instance.pages,
+          translations: LocalizationService(),
+          locale: LocalizationService.locale,
+          fallbackLocale: LocalizationService.fallbackLocale,
+          theme: theme,
+        ),
       ),
     );
   }